+++ src/components/Context/AuthContext.jsx
@@ -1,29 +1,6 @@
-/* eslint-disable no-restricted-syntax */
 import { createContext, useContext, useEffect, useMemo, useState } from "react";
-import {
-  createUserWithEmailAndPassword,
-  signInWithEmailAndPassword,
-  onAuthStateChanged,
-  signOut,
-  GoogleAuthProvider,
-  signInWithPopup,
-  updateProfile,
-  sendPasswordResetEmail,
-} from "firebase/auth";
-import {
-  doc,
-  setDoc,
-  collection,
-  getDocs,
-  getDoc,
-  updateDoc,
-  onSnapshot,
-} from "firebase/firestore";
-import { ref, uploadBytes, getDownloadURL } from "firebase/storage";
 import PropTypes from "prop-types";
-import Swal from "sweetalert2";
-import withReactContent from "sweetalert2-react-content";
-import { auth, fireStore, storage } from "../../utils/firebase/firebaseconfig";
+import { authUser } from "../../utils/auth";
 
 const authContext = createContext();
 
@@ -39,304 +16,18 @@ export function AuthProvider({ children }) {
   const [sessionUser, setSessionUser] = useState(null);
   const [loading, setLoading] = useState(true);
 
-  const MySwal = withReactContent(Swal);
   useEffect(() => {
-    onAuthStateChanged(auth, (currentUser) => {
-      setSessionUser(currentUser);
-      setLoading(false);
-    });
+    authUser(setSessionUser, setLoading);
   }, []);
 
-  const useSweetAlert = (title, message, tipeIcon) => {
-    MySwal.fire(title, message, tipeIcon);
-  };
+  const data = useMemo(
+    () => ({
+      sessionUser,
+      loading,
+    }),
+    [sessionUser, loading],
+  );
 
-  const signup = async (email, password) => {
-    await createUserWithEmailAndPassword(auth, email, password);
-  };
-
-  const login = async (email, password) => {
-    await signInWithEmailAndPassword(auth, email, password);
-  };
-
-  const logOut = async () => {
-    try {
-      return await signOut(auth);
-    } catch (error) {
-      return useSweetAlert("Fibase error", error.message, "error");
-    }
-  };
-
-  const googleLogin = async () => {
-    try {
-      const googleProvider = new GoogleAuthProvider();
-      return await signInWithPopup(auth, googleProvider);
-    } catch (error) {
-      return useSweetAlert("Fibase error", error.message, "error");
-    }
-  };
-
-  const updateUser = async () => {
-    try {
-      return await updateProfile(auth.currentUser, {
-        displayName: "ADMIN",
-        photoURL:
-          "https://thumbs.dreamstime.com/z/inicio-de-sesi%C3%B3n-administrador-en-el-icono-del-port%C3%A1til-vector-stock-166205404.jpg",
-      });
-    } catch (error) {
-      return useSweetAlert("Fibase error", error.message, "error");
-    }
-  };
-
-  const resetPassword = async (email) => {
-    try {
-      return await sendPasswordResetEmail(auth, email);
-    } catch (error) {
-      return useSweetAlert("Fibase error", error.message, "error");
-    }
-  };
-
-  const createUrbanization = async (data) => {
-    try {
-      const docRef = doc(fireStore, sessionUser.email, data.identification);
-      return await setDoc(docRef, data);
-    } catch (error) {
-      return useSweetAlert("Fibase error", error.message, "error");
-    }
-  };
-
-  const fileHandler = async (e, urbanizationName) => {
-    try {
-      const localFile = e.target.files[0];
-      const fileRef = ref(storage, `${urbanizationName}/${localFile.name}`);
-      await uploadBytes(fileRef, localFile);
-      const urlDown = await getDownloadURL(fileRef);
-      return urlDown;
-    } catch (error) {
-      return useSweetAlert("Fibase error", error.message, "error");
-    }
-  };
-
-  const getAllUrbanization = async () => {
-    try {
-      const docRef = await getDocs(collection(fireStore, sessionUser.email));
-      let infor = [];
-      docRef.forEach((docInf) => {
-        infor = [...infor, docInf.data()];
-      });
-      return infor;
-    } catch (error) {
-      return useSweetAlert("Fibase error", error.message, "error");
-    }
-  };
-
-  const addHouse = async (id, data) => {
-    const { housenumber } = data;
-    try {
-      const docRef = doc(fireStore, sessionUser.email, id.id);
-      const docSnap = await getDoc(docRef);
-      const { house } = docSnap.data();
-      await updateDoc(docRef, { house: { ...house, [housenumber]: data } });
-    } catch (error) {
-      useSweetAlert("Fibase error", error.message, "error");
-    }
-  };
-
-  const realtimeCollectionCheck = async (id, setHouse) => {
-    try {
-      const unsub = onSnapshot(
-        doc(fireStore, sessionUser.email, id.id),
-        (house) => {
-          setHouse(house.data());
-        },
-      );
-      return unsub;
-    } catch (error) {
-      return useSweetAlert("Fibase error", error.message, "error");
-    }
-  };
-  const realtimeCollectionVote = async (setVoteCollection, code) => {
-    try {
-      const unsub = onSnapshot(doc(fireStore, "vote", code), (vote) => {
-        setVoteCollection(vote.data());
-      });
-      return unsub;
-    } catch (error) {
-      return useSweetAlert("Fibase error", error.message, "error");
-    }
-  };
-
-  const updateActiveHouse = async (id, houseNumber) => {
-    try {
-      const docRef = doc(fireStore, sessionUser.email, id);
-      const docSnap = await getDoc(docRef);
-      const { house } = docSnap.data();
-      const { votacion } = house[houseNumber];
-      await updateDoc(docRef, {
-        house: {
-          ...house,
-          [houseNumber]: { ...house[houseNumber], votacion: !votacion },
-        },
-      });
-    } catch (error) {
-      useSweetAlert("Fibase error", error.message, "error");
-    }
-  };
-
-  const createVote = async (initialData, codevote) => {
-    try {
-      const docRef = doc(fireStore, "vote", codevote);
-      return await setDoc(docRef, initialData);
-    } catch (error) {
-      return useSweetAlert("Fibase error", error.message, "error");
-    }
-  };
-
-  const addquestion = async (codevote, questionToAdd) => {
-    const docRef = doc(fireStore, "vote", codevote);
-    const docData = await getDoc(docRef);
-    const { questions } = docData.data();
-    const newQuestion = {
-      [Object.values(questions).length + 1]: {
-        questionToAdd,
-        yes: [],
-        not: [],
-        state: true,
-      },
-    };
-    updateDoc(docRef, { questions: { ...questions, ...newQuestion } });
-  };
-
-  const checkVoteHouse = async (code) => {
-    try {
-      const docRef = doc(fireStore, "vote", code);
-      const docData = await getDoc(docRef);
-      if (docData.exists()) {
-        const data = docData.data();
-        return data;
-      }
-      return useSweetAlert(
-        "error",
-        "codigo de votacion no encontrado",
-        "error",
-      );
-    } catch (error) {
-      return useSweetAlert("error", error.message, "error");
-    }
-  };
-
-  const resetVoteActive = async (listVoteHouse, code) => {
-    try {
-      const { houseVoteActive, questions } = listVoteHouse;
-      let houseRestObj = {};
-
-      for (const restStateHouseVote in houseVoteActive) {
-        if (listVoteHouse) {
-          houseRestObj = {
-            ...houseRestObj,
-            [restStateHouseVote]: {
-              ...houseVoteActive[restStateHouseVote],
-              votacion: true,
-            },
-          };
-        }
-      }
-
-      const docRef = doc(fireStore, "vote", code);
-      return updateDoc(docRef, {
-        houseVoteActive: houseRestObj,
-        questions: {
-          ...questions,
-          [Object.keys(questions).length]: {
-            ...questions[Object.keys(questions).length],
-            state: false,
-          },
-        },
-      });
-    } catch (error) {
-      return useSweetAlert("error", error.message, "error");
-    }
-  };
-
-  const addVote = async (choice, house, code, listVoteHouse) => {
-    try {
-      const { questions, houseVoteActive } = listVoteHouse;
-
-      const docRef = doc(fireStore, "vote", code);
-      updateDoc(docRef, {
-        questions: {
-          ...questions,
-          [Object.keys(questions).length]: {
-            ...questions[Object.keys(questions).length],
-            [choice]:
-              questions[Object.keys(questions).length][choice].concat(house),
-          },
-        },
-        houseVoteActive: {
-          ...houseVoteActive,
-          [house]: { ...houseVoteActive[house], votacion: false },
-        },
-      });
-      return null;
-    } catch (error) {
-      return useSweetAlert("error", error.message, "error");
-    }
-  };
-
-  const endVote = async (code, id) => {
-    try {
-      const docRef = doc(fireStore, sessionUser.email, id);
-      const voteRef = doc(fireStore, "vote", code);
-      const docData = await getDoc(docRef);
-
-      const { house } = docData.data();
-
-      let houseRestObj = {};
-
-      for (const restStateHouseVote in house) {
-        if (house) {
-          houseRestObj = {
-            ...houseRestObj,
-            [restStateHouseVote]: {
-              ...house[restStateHouseVote],
-              votacion: false,
-            },
-          };
-        }
-      }
-      await updateDoc(voteRef, { state: false });
-      await updateDoc(docRef, { house: houseRestObj });
-      return null;
-    } catch (error) {
-      return useSweetAlert("error", error.message, "error");
-    }
-  };
-
-  const data = useMemo(() => ({
-    signup,
-    login,
-    endVote,
-    sessionUser,
-    logOut,
-    loading,
-    useSweetAlert,
-    googleLogin,
-    updateUser,
-    resetPassword,
-    createUrbanization,
-    fileHandler,
-    getAllUrbanization,
-
-    addHouse,
-    realtimeCollectionCheck,
-    createVote,
-    updateActiveHouse,
-    addquestion,
-    realtimeCollectionVote,
-    checkVoteHouse,
-    resetVoteActive,
-    addVote,
-  }));
   return <authContext.Provider value={data}>{children}</authContext.Provider>;
 }
