+++ src/components/ResultQuestion/index.jsx
@@ -1,8 +1,9 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { Bar } from "react-chartjs-2";
-// eslint-disable-next-line no-unused-vars
-import { Chart as ChartJS } from "chart.js/auto";
+import { Chart as ChartJS, registerables } from "chart.js";
+
+ChartJS.register(...registerables);
 
 function ResultQuestion({ dataVote }) {
   const { houseVoteActive, questions } = dataVote;
